added total controllers and error management for ciaClient delete controller

+++ controllers/ciaClient.controllers.js
@@ -5,7 +5,7 @@ CIA_Client.hasOne(Price, { foreignKey: "id" });
 
 const getCIAClient = async (req, res) => {
   try {
-    const client = await CIA_Client.findOne({
+    let client = await CIA_Client.findOne({
       where: { id: req.body.id },
       attributes: {
         exclude: ["id", "id_price"],
@@ -25,8 +25,9 @@ const getCIAClient = async (req, res) => {
         .status(404)
         .json({ success: false, message: "CIA Client could not be found." });
     } else {
-      client.dataValues.Price = client.dataValues.Price.dataValues;
-      res.status(200).json({ success: true, data: client.dataValues });
+      client = client.dataValues;
+      client.Price = client.Price.dataValues;
+      res.status(200).json({ success: true, data: client });
     }
   } catch (error) {
     res.status(400).json({ message: `ERROR: ${error.stack}` });
@@ -39,7 +40,6 @@ const createCIAClient = async (req, res) => {
     let data = req.body.client;
     data.id_price = price.dataValues.id;
     const client = await CIA_Client.create(data);
-    client.price = price.dataValues;
     res.status(201).json({ success: true, data: client.dataValues });
   } catch (error) {
     res.status(400).json({ message: `ERROR: ${error.stack}` });
@@ -47,20 +47,26 @@ const createCIAClient = async (req, res) => {
 };
 
 const deleteCIAClient = async (req, res) => {
-    try {
-        const client = await CIA_Client.findOne({where: {id: req.body.id}});
-        await Price.destroy({where: {id: client.dataValues.id_price}});
+  try {
+    const client = await CIA_Client.findOne({ where: { id: req.body.id } });
+    if (!client) {
+        res.status(404).json({success: false, message: `CIA client with id ${req.body.id} does not exist.`})
+    } else {
+        await Price.destroy({ where: { id: client.dataValues.id_price } });
         await client.destroy();
-        res.status(200).json({success: true, message: 'CIA Client deleted successfully.'});
-    } catch (error) {
-        res.status(400).json({ message: `ERROR: ${error.stack}` });
+        res
+          .status(200)
+          .json({ success: true, message: "CIA Client register deleted successfully." });
     }
+  } catch (error) {
+    res.status(400).json({ message: `ERROR: ${error.stack}` });
+  }
 };
 
 const controllers = {
-    getCIAClient,
-    createCIAClient,
-    deleteCIAClient
+  getCIAClient,
+  createCIAClient,
+  deleteCIAClient,
 };
 
 module.exports = controllers;
